# fix(api): optimise archived list query. Fixes #13295