@bom-d-van 2024-09-08
hi @agilgur5  thank for the feedback. I have pushed another fix for mysql. it requires a some extra handling.

there is till one more test failure, but it seems not related to my changes. do you have any suggestions? 
or is it safe to leave it as it is.