@agilgur5 2024-09-08
> there is a test failure about (`artgc-artifact-not-written-failed-zn9dv`), but it seems not related to my changes. do you have any suggestions?



> hmm, it seems that retry helps. :P

Sounds like a test flake then. Although it looks like it [was with `TestStoppedWorkflow`](https://github.com/argoproj/argo-workflows/actions/runs/10758518983/job/29834036494), and not the test you mentioned above (which was added recently):

```
artifacts_test.go:218: timeout after 1m30s waiting for condition
    artifacts_test.go:241: 
        	Error Trace:	/home/runner/work/argo-workflows/argo-workflows/test/e2e/artifacts_test.go:241
        	            				/home/runner/work/argo-workflows/argo-workflows/test/e2e/fixtures/then.go:251
        	            				/home/runner/work/argo-workflows/argo-workflows/test/e2e/artifacts_test.go:240
        	Error:      	Received unexpected error:
        	            	The specified key does not exist.
        	Test:       	TestArtifactsSuite/TestStoppedWorkflow
=== FAIL: ArtifactsSuite/TestStoppedWorkflow
```