@agilgur5 2024-09-08
> the query runtime on our production env is reduced from 11495.734 ms to 44.713 ms. the data size is about 417481 argo_archived_workflows and 1794624 argo_archived_workflows_labels.

While this requires a good bit of rows to test, I'm also thinking that we could potentially write a simple test that checks if the query returns under a certain amount of time on a test seed. For example, under 100ms.
That way it could act as a regression test for anything severely off, but not a full performance test to catch everything.