@bom-d-van 2024-09-09
> While this requires a good bit of rows to test, I'm also thinking that we could potentially write a simplified test that checks if the query returns under a certain amount of time on a test seed. For example, under 100ms.

@agilgur5  yeah, it sounds quite reasonable and would be helpful in the long run. but unfortunately, I didn't spot a simple way for me to add a new regression test fast enough and my recent work schedule is quite tight.

could we add a new issue and linking it to this pr for now? I might be able to make the contribution a few weeks later.