@agilgur5 2024-09-10
> could we add a new issue and linking it to this pr for now? I might be able to make the contribution a few weeks later.

Yea that'd be fine so long as we get some more manual tests of this. I'm loading up my dev env right now to push up an image