@ryancurrah 2024-09-10
Provide me an image to use and I will test it.