@agilgur5 2024-09-10
Pushed up [`agilgur5/argo-cli:fix-archive-list-query-optimization`](https://hub.docker.com/layers/agilgur5/argo-cli/fix-archive-list-query-optimization/images/sha256-d3a31a31e051a499e7b49e93a7c942414cf83eb56d6e46cc012b23758e1b36a7?context=repo) for testing