@ryancurrah 2024-09-12
> Thanks for the testing @ryancurrah !
> 
> With that, some of the individual query tests folks made, and 2 other approvals, I think this is good to go 🚀
> 
> Note that this query seems to be further optimizable per [#13295 (comment)](https://github.com/argoproj/argo-workflows/issues/13295#issuecomment-2340964248), so we may want to open a new issue to add further optimizations, but this at least fixes the performance issues and gets us to a much more reasonable baseline

If you need someone to test any more query optimizations feel free to ping me.