@agilgur5 2024-09-14
Added a follow-up issue for further optimizations in https://github.com/argoproj/argo-workflows/issues/13601